+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * `CollectionsPlume.Replacement`
   * `EntryReader.Entry`
 * New overloads of `UtilPlume.backticks()` that take a directory argument.
+* New method `SystemPlume.pathToExecutable()`.
 
 ## 1.14.0 (2026-03-05)